Move family_graph.js to webpack
* Remove cytoscape dependencies
