+++ core/templates/core/user_pictures.jinja
@@ -5,7 +5,7 @@
 {%- endblock -%}
 
 {% block additional_js %}
-  <script src="{{ static('webpack/users/pictures-index.js') }}" defer></script>
+  <script src="{{ static('webpack/user/pictures-index.js') }}" defer></script>
 {% endblock %}
 
 {% block title %}
