+++ package.json
@@ -31,6 +31,9 @@
     "@fortawesome/fontawesome-free": "^6.6.0",
     "@zip.js/zip.js": "^2.7.52",
     "alpinejs": "^3.14.1",
+    "cytoscape": "^3.30.2",
+    "cytoscape-cxtmenu": "^3.5.0",
+    "cytoscape-klay": "^3.1.4",
     "easymde": "^2.18.0",
     "glob": "^11.0.0",
     "jquery": "^3.7.1",
