+++ sith/settings.py
@@ -740,10 +740,6 @@ if SENTRY_DSN:
 SITH_FRONT_DEP_VERSIONS = {
     "https://github.com/chartjs/Chart.js/": "2.6.0",
     "https://github.com/getsentry/sentry-javascript/": "8.26.0",
-    "https://github.com/cytoscape/cytoscape.js": "3.30.2 ",
-    "https://github.com/cytoscape/cytoscape.js-cxtmenu": "3.5.0",
-    "https://github.com/cytoscape/cytoscape.js-klay": "3.1.4",
-    "https://github.com/kieler/klayjs": "0.4.1",  # Deprecated, elk should be used but cytoscape-elk is broken
     "https://github.com/mrdoob/three.js/": "r148",
     "https://github.com/vasturiano/three-spritetext": "1.6.5",
     "https://github.com/vasturiano/3d-force-graph/": "1.70.19",
